--- run_test/file2
@@ -1,1 +1,0 @@
-hello world
