+++ run_test/srcs/big_cutter.c
@@ -1,5 +1,3 @@
-
-
 #include "minishell.h"
 
 t_nd	*big_cutter(char *str)
@@ -22,7 +20,6 @@ t_nd	*big_cutter(char *str)
 		if (i == 0)
 		{
 			mother->child = new_nd(tmp[i]);
-			// mother->child->prev = mother;
 			tmp_nd = mother->child;
 		}
 		else
