+++ run_test/srcs/free_memory.c
@@ -15,33 +15,29 @@ void	free_nd_data(t_nd *coms)
 		free(coms->args);
 }
 
-void	free_sible(t_nd *coms)
+void	free_nd(t_nd *coms)
 {
-	t_nd *tmp;
-	t_nd *nd_erase;
+	t_nd	*tmp;
+	t_nd	*nd_erase;
 
 	tmp = coms;
 	while (tmp->sible)
 		tmp = tmp->sible;
 	if (tmp->child)
-		free_sible(tmp->child);
-	
+		free_nd(tmp->child);
 	while (tmp)
 	{
-		free_nd_data(tmp);
 		nd_erase = tmp;
+		free_nd_data(nd_erase);
 		if (tmp->prev)
+		{
 			tmp = tmp->prev;
-		free(nd_erase);
+			free(nd_erase);
+		}
+		else
+		{
+			free(tmp);
+			tmp = 0;
+		}
 	}
-}
-
-void	free_nd(t_nd *coms)
-{
-	t_nd *tmp;
-	
-	tmp = coms->child;
-	free_nd_data(coms);
-	free_sible(tmp);
-	free(coms);
 }
