+++ run_test/srcs/main.c
@@ -20,7 +20,7 @@ int		main(int ac, char *av[], char *en[])
 {
 	char	**env;
 	char	rt;
-	
+
 	g_ex.exit_code = 0;
 	init_env(&env, en);
 	rt = start_shell(&env, av[0]);
