+++ run_test/srcs/split_quote.c
@@ -24,9 +24,7 @@ int	count_size(char *str, char *set)
 			q_f *= -1;
 		else if (ft_strchr(set, str[i]) && qq_f > 0 && q_f > 0)
 		{
-			if (ft_strchr(set, str[i + 1]))
-				;
-			else
+			if (!ft_strchr(set, str[i + 1]))
 				count++;
 		}
 		i++;
@@ -56,9 +54,6 @@ char	*sep(char **str, char *set)
 			q_f *= -1;
 		if (ft_strchr(set, (*str)[i]) && qq_f > 0 && q_f > 0)
 		{
-			// if (ft_strchr(set, (*str)[i + 1]))
-			// 	;
-			// else
 			rt = ft_strndup(*str, i);
 			break ;
 		}
@@ -72,18 +67,6 @@ char	*sep(char **str, char *set)
 	while (ft_strchr(set, (*str)[i]))
 		i++;
 	(*str) += i;
-
-	// if (!(*str)[i])
-	// {
-	// 	if (ft_strchr(set, (*str)[i - 1]))
-	// 		i--;
-	// 	rt = ft_strndup(*str, i);
-	// }
-	// else
-	// {
-	// 	rt = ft_strndup(*str, i);
-	// 	(*str) += i + 1;
-	// }
 	return (rt);
 }
 
@@ -102,13 +85,6 @@ char	**split_quote(char *str, char *set)
 	i = -1;
 	while (++i < size)
 		rt[i] = sep(&str, set);
-
-	// j = ft_strlen(rt[i - 1]) - 1;
-	// while (rt[i - 1][j] == ' ')
-	// {
-	// 	j--;
-	// 	rt[i - 1][j] = 0;
-	// }
 	rt[i] = 0;
 	return (rt);
 }
